+++ src/8thWonderland/Application/Controller/DeveloperController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Wonderland\Application\Controller;
+
+use Wonderland\Library\Controller\ActionController;
+
+class DeveloperController extends ActionController {
+    public function installationGuideAction() {
+        return $this->render('developer/installation_guide');
+    }
+}